requirements.txt
Flask==3.1.2
flask-cors==5.0.0
gspread==6.2.1
google-auth==2.42.1
google-auth-oauthlib==1.2.2
python-dotenv==1.2.1
gunicorn==23.0.0
